+++ src/index.js
@@ -2,6 +2,8 @@ import { loadConfig } from "./config/config.js";
 import { validateConfig } from "./config/validate.js";
 import { MCPServer } from "./core/server.js";
 import { ping } from "./tools/ping.js";
+import { listDir } from "./tools/list_dir.js";
+import { openFile } from "./tools/open_file.js";
 
 const config = loadConfig();
 validateConfig(config);
@@ -9,7 +11,9 @@ validateConfig(config);
 const server = new MCPServer({
   config,
   tools: {
-    ping
+    ping,
+    list_dir: listDir,
+    open_file: openFile
   }
 });